Extracts TextWithLinks component for reusability
Moves the TextWithLinks component to a common directory
for use in multiple panels and avoids code duplication.
Updates styles path.

Relates to feature/lost_omens-109
